Clean up sections batching in `ComponentEncoder`
`ComponentEncoder` should not have its own sophisticated batching logic. Ideally we shift definitions normalization to `ComponentWatParser`. This ensures that normalization logic as separated from the encoder and keeps it clean and lean. In the long term we should use `enum NormalizedDefinition` for accumulation in `ComponentWatParser`, case of this enum store an index to a corresponding component definition in `ComponentDef` (now moved to a separate file).
